Introduce Brian Troutwine's slides.
This is a nearly complete draft of Brian Troutwine's slides for the
2015 LambdaConf in Boulder, CO. A copy of the
Attribution-NonCommercial-NoDerivatives 4.0 International license is
included with these slides and they are released under its terms.

The fonts solely licensed to me included in these slides are:

  Fairview
  Klinic Slab

They may _not_ be used for commercial purposes without prior
arrangements from the Lost Type Co-Op, found online at
http://www.losttype.com/. The photography included is a mixture of
fair use, public domain and my own work. My own photographs are
licensed under the same terms as the slides. They are, as of this
commit, the photographs backing slides:

  1, 21, 32, 73, 80 and 87

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>